Database structure, Database Controller, Logger and FileManagement
Database structure sets the different table, type and value of a json file.
Database Controller will manage the saving / loading and it will be the interface with other entity.
Logger write a log in a file with horodatage and type (Warning, error ...).
FileManagement allows to read, write and add a content in a file.